+++ maestro-cli/src/main/java/maestro/cli/command/LoginCommand.kt
@@ -1,8 +1,8 @@
 package maestro.cli.command
 
 import maestro.auth.ApiKey
-import maestro.cli.DisableAnsiMixin
-import maestro.cli.ShowHelpMixin
+import maestro.cli.mixin.DisableAnsiMixin
+import maestro.cli.mixin.ShowHelpMixin
 import maestro.cli.analytics.Analytics
 import maestro.cli.analytics.UserLoggedOutEvent
 import maestro.cli.api.ApiClient
