+++ maestro-cli/src/main/java/maestro/cli/command/RecordCommand.kt
@@ -22,13 +22,13 @@ package maestro.cli.command
 import kotlinx.coroutines.runBlocking
 import maestro.cli.App
 import maestro.cli.CliError
-import maestro.cli.AppleTeamIdMixin
-import maestro.cli.ConfigFileMixin
-import maestro.cli.DebugOutputMixin
-import maestro.cli.DeviceSelectionMixin
-import maestro.cli.DisableAnsiMixin
-import maestro.cli.EnvMixin
-import maestro.cli.ShowHelpMixin
+import maestro.cli.mixin.AppleTeamIdMixin
+import maestro.cli.mixin.ConfigFileMixin
+import maestro.cli.mixin.DebugOutputMixin
+import maestro.cli.mixin.DeviceSelectionMixin
+import maestro.cli.mixin.DisableAnsiMixin
+import maestro.cli.mixin.EnvMixin
+import maestro.cli.mixin.ShowHelpMixin
 import maestro.cli.analytics.Analytics
 import maestro.cli.analytics.RecordFinishedEvent
 import maestro.cli.analytics.RecordStartedEvent
