+++ maestro-cli/src/main/java/maestro/cli/command/TestCommand.kt
@@ -27,17 +27,17 @@ import kotlinx.coroutines.runBlocking
 import maestro.Maestro
 import maestro.cli.App
 import maestro.cli.CliError
-import maestro.cli.ApiClientMixin
-import maestro.cli.AppleTeamIdMixin
-import maestro.cli.ConfigFileMixin
-import maestro.cli.DebugOutputMixin
-import maestro.cli.DeviceSelectionMixin
-import maestro.cli.DisableAnsiMixin
-import maestro.cli.EnvMixin
-import maestro.cli.ReinstallDriverMixin
-import maestro.cli.ReportOutputMixin
-import maestro.cli.ShowHelpMixin
-import maestro.cli.TagFilterMixin
+import maestro.cli.mixin.ApiClientMixin
+import maestro.cli.mixin.AppleTeamIdMixin
+import maestro.cli.mixin.ConfigFileMixin
+import maestro.cli.mixin.DebugOutputMixin
+import maestro.cli.mixin.DeviceSelectionMixin
+import maestro.cli.mixin.DisableAnsiMixin
+import maestro.cli.mixin.EnvMixin
+import maestro.cli.mixin.ReinstallDriverMixin
+import maestro.cli.mixin.ReportOutputMixin
+import maestro.cli.mixin.ShowHelpMixin
+import maestro.cli.mixin.TagFilterMixin
 import maestro.cli.analytics.Analytics
 import maestro.cli.analytics.TestRunFailedEvent
 import maestro.cli.analytics.TestRunFinishedEvent
