+++ maestro-cli/src/main/java/maestro/cli/mixin/ConfigFileMixin.kt
@@ -1,4 +1,4 @@
-package maestro.cli
+package maestro.cli.mixin
 
 import picocli.CommandLine
 import java.io.File
