+++ maestro-cli/src/main/java/maestro/cli/mixin/ReportOutputMixin.kt
@@ -1,4 +1,4 @@
-package maestro.cli
+package maestro.cli.mixin
 
 import maestro.cli.report.ReportFormat
 import picocli.CommandLine
